refactor: move 'normalizePaymentAmount' from payments module to graphql
