+++ src/graphql/root/mutation/index.ts
@@ -1,0 +1,11 @@
+import { ExchangeCurrencyUnit, WalletCurrency } from "@domain/shared"
+
+export const normalizePaymentAmount = (
+  paymentAmount: PaymentAmount<WalletCurrency>,
+): PaymentAmountPayload<ExchangeCurrencyUnit> => ({
+  amount: Number(paymentAmount.amount),
+  currencyUnit:
+    paymentAmount.currency === WalletCurrency.Usd
+      ? ExchangeCurrencyUnit.Usd
+      : ExchangeCurrencyUnit.Btc,
+})
