+++ src/graphql/root/mutation/index.types.d.ts
@@ -1,0 +1,4 @@
+type PaymentAmountPayload<T extends ExchangeCurrencyUnit> = {
+  currencyUnit: T
+  amount: number
+}
