+++ src/graphql/root/mutation/ln-usd-invoice-fee-probe.ts
@@ -1,5 +1,4 @@
 import { Payments } from "@app"
-import { normalizePaymentAmount } from "@domain/payments"
 
 import { GT } from "@graphql/index"
 import WalletId from "@graphql/types/scalar/wallet-id"
@@ -7,6 +6,8 @@ import CentAmountPayload from "@graphql/types/payload/sat-amount"
 import LnPaymentRequest from "@graphql/types/scalar/ln-payment-request"
 import { mapError } from "@graphql/error-map"
 import { validateIsUsdWalletForMutation } from "@graphql/helpers"
+
+import { normalizePaymentAmount } from "."
 
 const LnUsdInvoiceFeeProbeInput = GT.Input({
   name: "LnUsdInvoiceFeeProbeInput",
